+++ CHANGELOG.md
@@ -5,10 +5,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## [0.0.4] - 2016-12-05
 ### Added
  - A seed file.
  - Routes for hangup action and direct call.
  - A task file to init plivo applications and endpoints.
+
+### Changed
+ - This file add wrong url for tags link.  
 
 ## [0.0.3] - 2016-12-05
 ### Added
@@ -40,6 +45,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
  - A .gitignore for rails app.
  - A Rails skeleton.
 
-[Unreleased]: https://github.com/olivierlacan/keep-a-changelog/compare/v0.0.3...HEAD
-[0.0.3]: https://github.com/olivierlacan/keep-a-changelog/compare/v0.0.2...v0.0.3
-[0.0.2]: https://github.com/olivierlacan/keep-a-changelog/compare/v0.0.1...v0.0.2
+[Unreleased]: https://github.com/Bit-Doctor/aircall-test/compare/v0.0.4...HEAD
+[0.0.4]: https://github.com/Bit-Doctor/aircall-test/compare/v0.0.3...v0.0.4
+[0.0.3]: https://github.com/Bit-Doctor/aircall-test/compare/v0.0.2...v0.0.3
+[0.0.2]: https://github.com/Bit-Doctor/aircall-test/compare/v0.0.1...v0.0.2